requirements.txt

matplotlib==3.3.3
numpy==1.19.2
opencv-python==4.4.0.46
PyYAML==5.3.1
scikit-learn==0.23.2
torch==1.8.0
tqdm==4.53.0
wandb==0.12.11

